requirements..txt
Flask
flask-cors
waitress
python-dotenv
google-generativeai
PyMuPDF
